refactored nft form files
